backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.2.0
python-slugify==8.0.1
httpx==0.26.0
alembic==1.13.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
